Feature/compodoc clean (#23)
Setup is done following the compodoc tutorial and using `ng add` schematics

* Documentation Enhancement: Compodoc

Compodoc is an angular extension that automatically generates a
documentation from the code with just one command.
Changes:
Compodoc initial setup via angular cli
Adding the documentation output to the .gitignore
Updating the ReadME on how to use the documentation

* Update README.md for compodoc

Updated the README with the link to compodoc.
Updated the description of the link to the documentation.
Updated the package.json with the initial setup for starting compodoc.
----
Co-authored-by: Tabea Röthemeyer <tabea.roethemeyer@gruppe.ai>